Progress on issue 108 - Big Rock Ending freestyle sections are now ACTIVE.  They are supposed to light up when you hit the lanes but they currently do not.
However, every note hit in a freestyle section rewards 5 points (I know this needs attention but it will do for now).  

It also seems to be a bit difficult to maintain your streak through to the end of a BRE; it seems fine keeping the streak on entry.  

Also, the notes currently still render on top of the BRE lanes for now.